Disable all default features of git2
We don't need the SSH or HTTPS features of git2 since we make no
modifications to the repository and do not push it anywhere. We just
fetch the metadata.

To shorten the build time and executable size, remove the features. This
excludes the OpenSSL library from compilation.

This also solves the additional dependency problem for NixOS.